drop cast for CURL_HTTP_VERSION_2TLS

+++ tests/libtest/lib1662.c
@@ -73,8 +73,7 @@ static CURLcode test_lib1662(char *URL)
       curl_easy_setopt(hnd, CURLOPT_USERAGENT, "curl/2000");
       curl_easy_setopt(hnd, CURLOPT_FOLLOWLOCATION, 1L);
       curl_easy_setopt(hnd, CURLOPT_MAXREDIRS, 50L);
-      curl_easy_setopt(hnd, CURLOPT_HTTP_VERSION,
-                       (long)CURL_HTTP_VERSION_2TLS);
+      curl_easy_setopt(hnd, CURLOPT_HTTP_VERSION, CURL_HTTP_VERSION_2TLS);
       curl_easy_setopt(hnd, CURLOPT_VERBOSE, 1L);
       curl_easy_setopt(hnd, CURLOPT_FTP_SKIP_PASV_IP, 1L);
       curl_easy_setopt(hnd, CURLOPT_TCP_KEEPALIVE, 1L);
